+++ waves/abaqus/_settings.py
@@ -1,0 +1,9 @@
+# Default supported file extensions
+_default_h5_extension = '.h5'
+_default_yaml_extension = '.yaml'
+_default_parsed_extension = '.parsed'
+_default_odbreport_extension = '.csv'
+_default_timestamp_format = '%Y%m%d-%H%M%S'
+_default_abaqus_command = 'abaqus'
+# Temporary contact info
+_parsing_error_contact = 'prabhu@lanl.gov'